python/requirements.txt
requests
tqdm
pypdf
langchain
langchain-openai
langchain-community
chromadb
fastapi
uvicorn
python-dotenv
